Major update: Unsupervised Causal Learning
This update covers the most intriguing ML part I have felt since starting CCNets.

A Cooperative Encoding Network consists of an Explainer, Reasoner, and Producer which is equal to components in Causal Cooperative Network, however, it doesn't need a label for its input, which can be categorized as unsupervised learning.

The Cooperative Encoding Network learns to factorize input into explanations extracted from Explainer and features which are the output of Reasoner. The explanation and feature represent deterministic and stochastic variables in probabilistic graphical models.

I realized we could learn the target prediction of observation in the dataset directly using the explainer's output. The decision model cannot learn any target prediction from the rest feature by Reasoner.

We need the feature paired with the explanation to reconstruct the original input.

I believe we really learn independent deterministic and stochastic variables, and the deterministic variable captures what we desired as target in any dataset. I will update on this further.